gopls: add WorkspaceFiles option
WorkspaceFiles allows an end-user to specify a set of files which, when
modified, will trigger a full reload of any views currently open in a
session. This is especially important for users who use custom
GOPACKAGESDRIVERS, as previously, you were forced to restart the
language server in order to get up-to-date diagnostics in some certain
instances.

Change-Id: Iba7a6137cb0b88a59318217a9a28d079100192a4